moved ad list to default home component
