+++ src/app/routes/home/default-home/components/ad-list/ad-list.component.scss
