+++ src/app/shared/components/connection-modal/connection-modal.component.ts
@@ -28,7 +28,6 @@ export class ConnectionModalComponent {
   isFormSubmitted(formSubmitted: boolean) {
     if (formSubmitted) {
       this.activeModal.close('Close click');
-      window.location.reload();
     }
   }
