+++ src/app/shared/components/connection-modal/login-form/login-form.component.ts
@@ -75,6 +75,7 @@ export class LoginFormComponent {
           if (data) {
             this.isFormSubmitted = true;
             this.formSubmitted.emit(this.isFormSubmitted);
+            window.location.reload()
           }
         },
         error: (error: any) => {
